Create transaction context
