+++ src/components/Summary/index.tsx
@@ -1,7 +1,11 @@
+import { useContext } from "react";
 import { SummaryCard, SummaryContainer } from "./styles";
 import { ArrowCircleUp, ArrowCircleDown, CurrencyDollar } from 'phosphor-react';
+import { TransactionContext } from "../../contexts/TransactionsContext";
 
 export function Summary() {
+  const { transactions } = useContext(TransactionContext)
+  
   return (
     <SummaryContainer>
       <SummaryCard>
