+++ src/pages/Transactions/index.tsx
@@ -1,35 +1,16 @@
-import { useEffect, useState } from "react";
+import { useContext } from "react";
 import { Header } from "../../components/Header";
 import { Summary } from "../../components/Summary";
 import { SearchForm } from "./components/SearchForm";
 import { 
   PriceHighlight, 
-  TransactionsContainer, 
+  TransactionsContainer,
   TransactionsTable 
 } from "./styles";
-
-interface Transaction {
-  id: number;
-  description: string; 
-  type: 'income' | 'outcome';
-  category: string;
-  price: number;
-  createdAt: string;
-}
+import { TransactionContext } from "../../contexts/TransactionsContext";
 
 export function Transactions() {
-  const [transactions, setTransactions] = useState<Transaction[]>([]);
-
-  async function loadTransactions() {
-    const response = await fetch('http://localhost:3333/transactions');
-    const data = await response.json();
-
-    setTransactions(data);
-  }
-
-  useEffect(() => {
-    loadTransactions();
-  }, []);
+  const { transactions } = useContext(TransactionContext)
 
   return (
     <div>
